Read proxy implementation ABI
Today it is hard to interact with proxy contracts on Celo Safe
because only the proxy contract ABI is loaded from Blockscout.

This PR performs a simple check for each address provided in the
"contract interaction" by performing a `getStorageAt` call for
the implementation slot (as defined in EIP-1967). If the address
is a valid contract address, then we attempt to fetch that ABI
as well and merge it with the proxy contract ABI.

Try the "contract interaction" flow with a proxy contract address.
Should be able to execute transaction on methods on the implementation
contract ABI.

n/a